[#18] Refactor: 패키지명 변경
