+++ app/src/main/java/com/example/cloop/data/model/cloth/ImageUploadResponse.kt
@@ -1,4 +1,4 @@
-package com.example.cloop.data.model.clothes
+package com.example.cloop.data.model.cloth
 
 data class ImageUploadResponse(
     val imageUrl: String
